requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via starlette
build==1.2.2.post1
    # via pip-tools
click==8.1.7
    # via
    #   pip-tools
    #   uvicorn
colorama==0.4.6
    # via
    #   build
    #   click
    #   pytest
dnspython==2.7.0
    # via pymongo
fastapi==0.115.4
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.10
    # via anyio
iniconfig==2.0.0
    # via pytest
motor[srv]==3.6.0
    # via -r requirements.in
packaging==24.1
    # via
    #   build
    #   pytest
pip-tools==7.4.1
    # via -r requirements.in
pluggy==1.5.0
    # via pytest
pydantic==2.9.2
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.23.4
    # via pydantic
pymongo==4.9.2
    # via motor
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.3.3
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
starlette==0.41.2
    # via fastapi
typing-extensions==4.12.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
uvicorn==0.32.0
    # via -r requirements.in
wheel==0.44.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
